+++ src/fs/block.h
@@ -20,5 +20,5 @@
 
 #include <stdbool.h>
 
-uint8_t* vfs_block_read(uint32_t offset, size_t size, uint8_t* buf);
-bool vfs_block_write(uint32_t offset, size_t size, uint8_t* buf);
+uint8_t* vfs_block_read(uint64_t offset, uint64_t size, uint8_t* buf);
+bool vfs_block_write(uint64_t offset, uint64_t size, uint8_t* buf);
